+++ db/migrate/20180222202201_create_anticipos.rb
@@ -4,10 +4,10 @@ class CreateAnticipos < ActiveRecord::Migration[5.1]
       t.references :senal, foreign_key: true, null:false
       t.integer :factura_id, null:false
       t.char :prefijo, limit: 6, null:false
-      t.decimal :nrofact, null:false
+      t.integer :nrofact, null:false
       t.integer :pago_id, null:false
       t.references :documento, foreign_key: true, null:false
-      t.decimal :nropago, null:false
+      t.integer :nropago, null:false
       t.datetime :fechatrn, null:false
       t.datetime :fechaven, null:false
       t.money :valor, null:false
